Création d'un UserControl pour hoster la fenetre OpenGl. Affichage et redimensionnement de la fenetre. Aucun évent géré pour le moment.
refs #1086 @4h00 &Affichage dans un controle pour WPF qui prend toute la place pour le moment&

+++ trunk/Commun/Utilitaire/Reseau/PaquetHandlers/PacketHandlerBase.cpp
@@ -21,7 +21,7 @@ void PacketHandlerBase::handlePacketPreparationBase(Paquet* pPaquet, PacketBuild
 int PacketHandlerBase::getPacketSize(Paquet* pPaquet) const
 {
     // Size de Paquet
-    int mTaillePaquet = Paquet::sequenceIdentification.length()+1 + GestionnaireReseau::longueurMaxOperationReseau;
+    int mTaillePaquet = (int)(Paquet::sequenceIdentification.length()+1 + GestionnaireReseau::longueurMaxOperationReseau);
     // Size de PaquetBase
     mTaillePaquet += sizeof(PaquetBase::CompteurNumeroPaquet) + sizeof(mTaillePaquet);
     return mTaillePaquet; // TODO: REMOVE HARDCODED VALUE
